Fix Biome formatting; move composer serialization into utils
The quality check failed on formatting in the new test file. Also moves docToContent/contentToDoc out of the component file into the mentionComposer utils (with their tests) to address React Doctor's only-export-components warnings.

Generated-By: PostHog Code
Task-Id: 0c3b84d6-4b62-449f-9154-917103a1109a
