+++ packages/ui/src/features/canvas/utils/mentionComposer.test.ts
@@ -1,6 +1,14 @@
 import type { UserBasic } from "@posthog/shared/domain-types";
+import { getSchema } from "@tiptap/core";
+import Mention from "@tiptap/extension-mention";
+import { Node as PmNode } from "@tiptap/pm/model";
+import StarterKit from "@tiptap/starter-kit";
 import { describe, expect, it } from "vitest";
-import { filterMentionCandidates } from "./mentionComposer";
+import {
+  contentToDoc,
+  docToContent,
+  filterMentionCandidates,
+} from "./mentionComposer";
 
 function member(overrides: Partial<UserBasic> & { email: string }): UserBasic {
   return {
@@ -57,5 +65,55 @@ describe("filterMentionCandidates", () => {
 
   it("returns empty when nothing matches", () => {
     expect(filterMentionCandidates(members, "zzz")).toEqual([]);
+  });
+});
+
+describe("contentToDoc / docToContent", () => {
+  const schema = getSchema([StarterKit, Mention]);
+
+  function roundTrip(content: string): string {
+    return docToContent(PmNode.fromJSON(schema, contentToDoc(content)));
+  }
+
+  it.each([
+    ["plain text", "hello there"],
+    ["empty content", ""],
+    ["mention token", "hey @[Raquel Smith](raquel@posthog.com) hi"],
+    [
+      "multiple mentions",
+      "@[Ann Lee](ann@posthog.com) @[Bob Stone](bob@posthog.com)",
+    ],
+    ["multi-line", "first\nsecond\n\nfourth"],
+    ["mention across lines", "cc @[Ann Lee](ann@posthog.com)\nthanks"],
+  ])("round-trips %s", (_label, content) => {
+    expect(roundTrip(content)).toBe(content);
+  });
+
+  it("maps mention tokens to mention nodes with email as id", () => {
+    const doc = contentToDoc("hi @[Ann Lee](ann@posthog.com)");
+    expect(doc.content?.[0]?.content).toEqual([
+      { type: "text", text: "hi " },
+      {
+        type: "mention",
+        attrs: { id: "ann@posthog.com", label: "Ann Lee" },
+      },
+    ]);
+  });
+
+  it("serializes hard breaks as newlines", () => {
+    const doc = PmNode.fromJSON(schema, {
+      type: "doc",
+      content: [
+        {
+          type: "paragraph",
+          content: [
+            { type: "text", text: "one" },
+            { type: "hardBreak" },
+            { type: "text", text: "two" },
+          ],
+        },
+      ],
+    });
+    expect(docToContent(doc)).toBe("one\ntwo");
   });
 });
